Put Mass copy Utility in place

+++ app/models/.svn/text-base/organization.rb.svn-base
@@ -731,16 +731,19 @@ class Organization < ActiveRecord::Base
      end
   end
 
-  def elt_reset_org_option
-    if self.elt_org_option
-      self.elt_org_option.destroy
+  def reset_org_option(app)
+    if app.elt?
+      if self.elt_org_option
+        self.elt_org_option.destroy
+      end
+      options = EltOrgOption.new
+      options.organization_id = self.id
+      options.owner_org_id = CoopApp.elt.first.owner.id
+      options.elt_cycle_id = nil
+      options.elt_framework_id = nil
+      self.elt_org_option = options
     end
-    options = EltOrgOption.new
-    options.organization_id = self.id
-    options.owner_org_id = CoopApp.elt.first.owner.id
-    options.elt_cycle_id = nil
-    options.elt_framework_id = nil
-    self.elt_org_option = options
+
   end
     
   def ifa_remove_org_options
